// cape-sdk/cape-sdk-angular/projects/cape-sdk-angular/src/lib/model/service-link/surrogateIdResponse.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface SurrogateIdResponse {
  surrogate_id?: string;
}