// cape-sdk/cape-sdk-angular/projects/cape-sdk-angular/src/lib/model/service-link/startLinkingRequest.ts
/**
 * SDK Service API
 * SDK Service API for integration with cape
 *
 * The version of the OpenAPI document: 2.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface StartLinkingRequest {
  session_code: string;
  surrogate_id: string;
  operator_id: string;
  service_id: string;
  return_url: string;
}